requirements.txt
Flask
flask-cors
openai
PyMuPDF
python-dotenv
